+++ vck/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/SdJwtCreatorTest.kt
@@ -7,6 +7,8 @@ import at.asitplus.wallet.lib.data.CredentialToJsonConverter.toJsonElement
 import at.asitplus.wallet.lib.data.SdJwtConstants
 import com.benasher44.uuid.uuid4
 import de.infix.testBalloon.framework.core.testSuite
+import io.kotest.assertions.withClue
+import io.kotest.matchers.collections.shouldBeSingleton
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.nulls.shouldBeNull
 import io.kotest.matchers.nulls.shouldNotBeNull
@@ -108,17 +110,19 @@ val SdJwtCreatorTest by testSuite {
 
     "array not selectively disclosable, but elements within" {
         listOf(
-            ClaimToBeIssued("array", listOf(
-                ClaimToBeIssuedArrayElement("1", true),
-                ClaimToBeIssuedArrayElement("2", false)
-            ), false)
+            ClaimToBeIssued(
+                "array", listOf(
+                    ClaimToBeIssuedArrayElement("1", false),
+                    ClaimToBeIssuedArrayElement("2", true)
+                ), false
+            )
         ).toSdJsonObject().apply {
             second.shouldHaveSize(1)
             first["array"].shouldNotBeNull().jsonArray.apply {
                 shouldHaveSize(2)
                 first() shouldBe JsonPrimitive("1")
                 get(1).jsonObject.shouldNotBeNull().apply {
-                    shouldHaveSize(1)
+                    entries.shouldBeSingleton()
                     get("...").shouldNotBeNull()
                 }
             }
@@ -127,10 +131,12 @@ val SdJwtCreatorTest by testSuite {
 
     "array selectively disclosable, and elements within too" {
         listOf(
-            ClaimToBeIssued("array", listOf(
-                ClaimToBeIssuedArrayElement("1", true),
-                ClaimToBeIssuedArrayElement("2", true)
-            ), true)
+            ClaimToBeIssued(
+                "array", listOf(
+                    ClaimToBeIssuedArrayElement("1", true),
+                    ClaimToBeIssuedArrayElement("2", true)
+                ), true
+            )
         ).toSdJsonObject().apply {
             second.shouldHaveSize(3)
             first["array"] shouldBe null